Refactor Belarus proxy endpoints into dedicated router
